d/postinst: order our reload override before the debhelper one
On fresh installation the service was first started, if enabled, and
only then it was enabled. While this does not matter for an
installation through the ISO, as there we don't start any services
anyway and after reboot in the installed system the service was
enabled and thus correctly started, it caused errors when using the
alternative of installing Proxmox VE on top of a plain Debian system.
As there, we currently recommend to reboot into the proxmox-kernel
first, to ensure all our required kernel features are available, and
only then direct the installation of Proxmox VE's user-space.

Naturally one could just recommend another reboot, but this here is
actually out of the line to all of our other postinst orders, so it
makes sense to align again even if we'd start recommending an extra
reboot after user-space installation.

Note that we only have to override here in the first place due to the
lack of reload support in the debhelpers, which we try to get in since
over 5 years [0].

[0]: https://salsa.debian.org/debian/debhelper/-/merge_requests/14

Fixes: https://forum.proxmox.com/threads/129901/
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>